% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nrw_api_meta}
\alias{nrw_api_meta}
\title{Return all available metadata}
\usage{
nrw_api_meta(api_key)
}
\arguments{
\item{api_key}{String. NRW Telemetry API key. String.}
}
\value{
list of length 5 with data.frames of all metadata queries
}
\description{
Returns a list with the following metadata queries: measurement areas, station statuses, operating areas, local authorities and measure types.
}
