% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nrw_api_request}
\alias{nrw_api_request}
\title{NRW Telemetry API Request for custom endpoints and parameters}
\usage{
nrw_api_request(endpoint, api_key, query_params = list())
}
\arguments{
\item{endpoint}{String for specific endpoint of the api.}

\item{api_key}{NRW Telemetry API key. String.}

\item{query_params}{A list of key-value pairs for query parameters. Parameters must be strings.}
}
\value{
Parsed JSON response from the API
}
\description{
This function sends a GET request to the NRW Telemetry API with support for query parameters and error handling.
}
