% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nrw_measure_values}
\alias{nrw_measure_values}
\title{Return Values for a Specific Measure}
\usage{
nrw_measure_values(api_key, measure, start = NULL, end = NULL)
}
\arguments{
\item{api_key}{NRW Telemetry API key. String.}

\item{measure}{String. Measure_id as given by nrw_station_measures.}

\item{start}{Optional. Date/Datetime object. Start of period.}

\item{end}{Optional. Date/Datetime object. End of period, leave blank to use current datetime.}
}
\value{
data.frame with measureId, quality code, timestamp, value
}
\description{
Returns values for a specific measure code. A time period can be provided to narrow the response. If no start or end date is provided then the default period is the preceding 24 hours from last value.
}
