% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nrw_station_measures}
\alias{nrw_station_measures}
\title{Retrieve Measures for a Given Station}
\usage{
nrw_station_measures(api_key, station, type = NULL)
}
\arguments{
\item{api_key}{NRW Telemetry API key. String.}

\item{station}{10 digit monitoring station reference code e.g H059000001 for Ynystanglws River Station. String.}

\item{type}{Station type, leave blank for measures of any type. River Level, Flow, Rainfall}
}
\value{
A data.frame where each row represents a different measure for a given station.
}
\description{
Retrieve measures for a given station and station type. The Measure_id can then be used to retreive values using nrw_measure_values.
}
