% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nrw_stations}
\alias{nrw_stations}
\title{Function to Return Stations based on Query Parameters}
\usage{
nrw_stations(
  api_key,
  stationName = NULL,
  riverName = NULL,
  hydrometricAreaId = NULL,
  localAuthority = NULL,
  nrwArea = NULL,
  floodTargetArea = NULL,
  status = NULL
)
}
\arguments{
\item{api_key}{NRW Telemetry API key. String.}

\item{stationName}{String. Partial match of all stations.}

\item{riverName}{String. Partial match of all river names.}

\item{hydrometricAreaId}{String. Hydrometric area ID as defined by CEH.}

\item{localAuthority}{String. Exact match for name of local authority.}

\item{nrwArea}{String. NRW Area.}

\item{floodTargetArea}{String. NRW Flood area.}

\item{status}{String. Exact match.}
}
\value{
data.frame containing matching stations and associated metadata.
}
\description{
Returns stations matching any of the provided query parameters.
}
